more updates, reintroduce menu buttons
