+++ app/views/pages/case/documents.html
@@ -28,8 +28,8 @@
 {% block header %}
   {{ super() }}
   {% set includeCaseActions = true %}
-
   {% include "includes/components/case/case-bar.html" %}
+
 
 {% endblock %}
