+++ app/views/pages/case/overview.html
@@ -53,7 +53,7 @@
 
 {% block header %}
   {{ super() }}
-  {% set includeCaseActions = false %}
+  {% set includeCaseActions = true %}
 
   {% include "includes/components/case/case-bar.html" %}
