+++ app/views/pages/case/products.html
@@ -28,7 +28,7 @@
 
 {% block header %}
   {{ super() }}
-  {% set includeCaseActions = false %}
+  {% set includeCaseActions = true %}
 
   {% include "includes/components/case/case-bar.html" %}
