+++ app/views/pages/case/timeline.html
@@ -45,7 +45,7 @@
 
 {% include "includes/components/case/case-navigation.html" %}
 
-<div class="govuk-grid-row">
+{# <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
 
     <h1 class="govuk-heading-l">Timeline</h1>
@@ -81,7 +81,7 @@
         </div>
       </div>
   </div>
-</div>
+</div> #}
 
 {{ activityTimeline( currentCase.activities, currentCase, currentTeam) }}
